api/requirements.txt
numpy
pillow
fastapi
uvicorn
tensorflow 
matplotlib
python-multipart
tensorflow-serving-api